+++ extensions/spotify-player/src/api/addToMySavedTracks.ts
@@ -13,6 +13,7 @@ export async function addToMySavedTracks({ trackIds }: AddToMySavedTracksProps) 
     const response = await spotifyClient.putMeTracks(trackIds.join());
     invalidateCache("api:library:tracks");
     invalidateCache("api:currently-playing");
+    invalidateCache(`api:liked:${trackIds.join(",")}`);
     return response;
   } catch (err) {
     const error = getErrorMessage(err);