+++ extensions/spotify-player/src/api/removeFromMySavedTracks.ts
@@ -13,6 +13,7 @@ export async function removeFromMySavedTracks({ trackIds }: RemoveFromMySavedTra
     const response = await spotifyClient.deleteMeTracks(trackIds.join());
     invalidateCache("api:library:tracks");
     invalidateCache("api:currently-playing");
+    invalidateCache(`api:liked:${trackIds.join(",")}`);
     return response;
   } catch (err) {
     const error = getErrorMessage(err);